Pass unfiltered compile comparisons to compile filters
So that they can correctly determine all found frontend thread values.
