+++ site/frontend/src/pages/compare/compile/filters.vue
@@ -5,13 +5,17 @@ import {computed, ref, toRaw, watch} from "vue";
 import {deepCopy} from "../../../utils/copy";
 import {PREF_FILTERS_OPENED} from "../prefs";
 import {createPersistedRef} from "../../../storage";
-import {CompileBenchmarkFilter, CompileTestCase, Target} from "./common";
-import {BenchmarkInfo, DEFAULT_FRONTEND_THREAD_COUNT} from "../../../api";
-import {TestCaseComparison} from "../data";
+import {
+  availableFrontendThreadsValues,
+  CompileBenchmarkFilter,
+  Target,
+} from "./common";
+import {BenchmarkInfo} from "../../../api";
+import {CompareResponse} from "../types";
 
 const props = defineProps<{
   info: BenchmarkInfo;
-  allComparisons: TestCaseComparison<CompileTestCase>[];
+  data: CompareResponse;
   // When reset, set filter to this value
   defaultFilter: CompileBenchmarkFilter;
   // Initialize the filter with this value
@@ -59,19 +63,9 @@ function updateSelfCompareParameter(event: Event) {
   filter.value.selfCompareParameter = rawValue === "" ? null : rawValue;
 }
 
-const availableFrontendThreadsValues = computed((): string[] => {
-  if (props.allComparisons.length === 0) {
-    return [DEFAULT_FRONTEND_THREAD_COUNT];
-  }
-  const uniqueFrontendThreads = [
-    ...new Set(props.allComparisons.map((c) => c.testCase.frontend_threads)),
-  ];
-  // Compare the string values as numbers
-  uniqueFrontendThreads.sort((a, b) =>
-    a.localeCompare(b, undefined, {numeric: true})
-  );
-  return uniqueFrontendThreads;
-});
+const frontendThreads = computed((): string[] =>
+  availableFrontendThreadsValues(props.data)
+);
 
 let filter = ref(deepCopy(props.initialFilter));
 watch(
@@ -280,7 +274,7 @@ const opened = createPersistedRef(PREF_FILTERS_OPENED);
               </div>
             </div>
             <ul class="states-list">
-              <li v-for="count in availableFrontendThreadsValues" :id="count">
+              <li v-for="count in frontendThreads" :id="count">
                 <label>
                   <input
                     type="checkbox"
